requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
transformers>=4.30.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP and Text Processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
emoji>=2.8.0
langdetect>=1.0.9

# Data Processing
datasets>=2.14.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
aiofiles>=23.0.0

# Model Evaluation
evaluate>=0.4.0
wandb>=0.15.0

# Deployment
gunicorn>=21.0.0
docker>=6.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
